Improve thread handling
Added a very simple lifeguard if the encoder is too slow: If the input
greatly outscales the output then the main thread is forced to wait for
the output to complete its current queue. This is not an issue on
presets >= medium however, which is most likely what you should be using
anyway  since this is realtime.