Fix for very misleading ffmpeg documentation
AVStream member "codec" (of type AVCodecContext) is marked as
deprecated, and it's suggested that you create your own AVCodecContext.
However, the "codec" variable is still the one that is used no matter.
In the header file this region is marked with "#if FF_API_LAVF_AVCTX",
if you recompile with this set to zero, you will get compilation errors
as other parts of ffmpeg references this "codec" variable. So in the
end, using the deprecated api is the way to go since they haven't
transitioned yet, but still marked it as deprecated?