Create game config installer
As current archive releases include a fixed version of the game config,
newer game additions will not be included in it. This installer will
supersede the reason to include the files in the release archives. The
user is intended to run this prior to loading.

The content downloaded is the latest available. The file
"Version\GameConfigLatest" was removed for redundancy and in favor of
"Output\SDR\GameConfigLatest".