Create SvrLockedArray & SvrLockedQueue
The recycled buffers do not need order, so those are replaced with just an array where items are pushed and popped from the back.

The lock free queues were bounded which opens the possibility of queue overflow when the writer keeps on being slightly faster than the reader. This is hard to solve, and even if solved by introducing blocks, you now have a bottleneck problem instead.

Order is very important in the audio / video queues so instead just grow additionally as long as we need to do so. The things we store are not particularly large and we have enough address space to store it.

This replaces the lock free queues with SvrLockedArray & SvrLockedQueue.