Create svr_log.dll and svr_console.dll
Needed to have svr_game.dll and svr_standalone.dll print to the same file.

A bit of a hack needed for svr_init in svr_game.dll to see if we are in integrated mode or standalone mode. In standalone mode, the log will already have been opened by svr_standalone.dll, so for integrated mode we have to create the log.

Also since svr_encoder.exe is only x64, but the other projects have either x86 or x64, svr_encoder.exe has a pre-build step that builds svr_log.dll in x64 so that it links properly. MSBuild does not have any way to retarget project references so it is inserted as a pre-build step.